Goodness of fit tests and plots
Page 1: For every GLMM we computed, we have included the output from RStudio’s function summary() applied to the model in question. Of special importance for goodness of fit are the following parameters: 

Akaike Information Criterion (AIC): Big negative numbers indicate good fit (relative)
Bayesian Information Criterion (BIC): Big negative numbers indicate good fit (relative)
Log-likelihood (logLik): Big positive numbers indicate the model explains well the data
Deviance: Big negative numbers = difference between fitted and observed values is low

Precision (ϕ) for beta-family models (PLVs): Big positive numbers indicate data is close to the mean (data is tightly clustered)
Dispersion (σ²) for gaussian (PLV differences) and gamma models (PTEs): Big negative numbers indicate low model variance (less spread in the residuals)

Conditional model: Significant fixed effects indicate those predictions explain a significant amount of variance after factoring out the random effects structure.

Page 2: For every GLMM we computed, we have included two plots from the RStudio DHARMa function, as well as a plot of the fitted versus the actual values. Of special importance for goodness of fit are the following points:

Plot 1.1 (top left) is a QQ plot which shows the distribution of plotted versus expected residuals across the entire spectrum of datapoints. Please note that deviations from normality for residuals are to be expected in real life data. This is especially the case in data such as ours, which contains outliers we decided not to remove. Moreover, data with non-gaussian distributions (beta & gamma) have a higher risk of displaying residuals which are skewed or that have non-constant variance. This is not an indication of bad fit.

Plot 1.2 (top left) is a variance plot, with measures for both intra- as well as inter-group variability. We note that Levene’s homogeneity of variance tests is non significant for most cases where DHARMa computes it (PTE values). This indicates that most models display homoscedasticity, which is of course desirable.

Plot 2 (top right) is a distribution plot of the standard deviations of the residuals (fitted vs observed values) with respect to simulated values. Most of the models display a non-significant deviation from the distribution, indicating that over- or under- dispersion is not a problem in the data. Some models do display such patterns, but we reiterate that this is to be expected for real world data and not a problem if the model otherwise fits well.  

Plot 3 (bottom) is a scatterplot of the observed versus the predicted values. Please note that the fit is tight for all models without exception, in alignment with values in Page 1.