Supplementary tables 1 & 2
Supplementary table 1 includes the percentage of variance explained across subjects and hemispheres for the Principal Component 1 (PC1) of each ROI used for PTE analyses (A1, dPMC, M1, vPMC). 

Supplementary table 2 contains the parts of the output from RStudio’s function ‘summary()’ that relate to goodness of fit, as applied to the model in question. Namely, the scalar values from the parameters below are collected in table form: 

-	Akaike Information Criterion (AIC): Big negative numbers indicate good fit (relative)
-	Bayesian Information Criterion (BIC): Big negative numbers indicate good fit (relative)
-	Log-likelihood (logLik): Big positive numbers indicate the model explains well the data
-	Deviance: Big negative numbers = difference between fitted and observed values is low

EITHER:
-	Precision (ϕ) for beta-family models (PLVs): Big positive numbers indicate data is close to the mean (data is tightly clustered)
OR:
-	Dispersion (σ²) for gaussian (PLV differences) and gamma models (PTEs): Big negative numbers indicate low model variance (less spread in the residuals)